+++ src/webgis4u/util/ol.js
@@ -1,0 +1,6 @@
+/**
+ * @fileoverview contains various util functions related to OpenLayers
+ * @module webgis4u/util/ol
+ */
+
+export * from './ol/getMapChildElementByClassName';